Add killfeed and in-game wishlist support
Adds live killfeed and in-game wishlist integration plus loot rendering/filter improvements.

- Config: new killfeed options (toggle, max entries, TTL, overlay pos) and additional loot UI settings (dot size, label font, height arrows, wishlist group toggles).
- Killfeed: new KillfeedEntry and KillfeedManager (thread-safe snapshot) with ImGui KillfeedPanel; kill events are pushed from LootManager when reading corpse dogtags.
- Wishlist: new WishlistManager reads the in-game WishlistManager from the local profile and exposes BSG ID → group lookup; LocalGameWorld and Memory now surface the wishlist manager and rate-limit refreshes.
- Loot/filter/render: LootFilter now considers in-game wishlist groups and exposes value tiers; LootItem drawing updated for tier-based sizing, halo rings, optional up/down height arrows and configurable label fonts; weapon names from dogtags are resolved to readable names when possible.
- Integration: LootManager retries partially-populated dogtags to reliably seed victims and push killfeed events; added retry caps and dedupe logic for pushes.
- Offsets: added wishlist-related offsets; small fixes across codebase (nullable out params, clearer ArgumentOutOfRange, QuestPlanner hash narrowing, Memory/Vmm arg list handling).

Also includes assorted UI/web updates and minor robustness fixes related to explosives and memory handling.

+++ src-silk/Tarkov/QuestPlanner/QuestPlannerWorker.cs
@@ -211,7 +211,7 @@ namespace eft_dma_radar.Silk.Tarkov.QuestPlanner
             for (int i = 0; i < list.Count; i++)
             {
                 var q = list[i];
-                ulong qfp = Mix((ulong)q.Id.GetHashCode() | ((ulong)statusTag << 56));
+                ulong qfp = Mix((ulong)(uint)q.Id.GetHashCode() | ((ulong)(uint)statusTag << 56));
                 foreach (var cond in q.CompletedConditions)
                     qfp ^= Mix((ulong)cond.GetHashCode());
                 bucket ^= qfp;
